# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Recipe API",
    "version": "1.0.0",
    "description": "API pentru gestionarea rețetelor culinare"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
"/api/interactions/favorites/{userId}": {
  "get": {
    "summary": "Obține rețetele favorite ale unui utilizator",
    "description": "Returnează lista rețetelor favorite pentru un utilizator specific",
    "parameters": [
      {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "ID-ul utilizatorului"
      }
    ],
    "responses": {
      "200": {
        "description": "Lista rețetelor favorite",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "ingredients": { "type": "string" },
                  "instructions": { "type": "string" },
                  "prepTime": { "type": "integer" },
                  "servings": { "type": "integer" },
                  "cuisineId": { "type": "integer" },
                  "image": { "type": "string" },
                  "userId": { "type": "integer" },
                  "createdAt": { "type": "string", "format": "date-time" },
                  "updatedAt": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
},
"/api/interactions/favorites/{recipeId}": {
  "post": {
    "summary": "Adaugă rețetă la favorite",
    "description": "Adaugă o rețetă la lista de favorite a utilizatorului autentificat",
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "recipeId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "ID-ul rețetei"
      }
    ],
    "responses": {
      "201": {
        "description": "Rețeta a fost adăugată la favorite",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "200": {
        "description": "Rețeta este deja la favorite",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "404": {
        "description": "Rețeta nu a fost găsită",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare la adăugarea la favorite",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  },
  "delete": {
    "summary": "Șterge rețetă din favorite",
    "description": "Elimină o rețetă din lista de favorite a utilizatorului autentificat",
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "recipeId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "ID-ul rețetei"
      }
    ],
    "responses": {
      "200": {
        "description": "Rețeta a fost eliminată din favorite",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare la eliminarea din favorite",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
},
"/api/interactions/follow/{userId}": {
  "post": {
    "summary": "Urmărește un utilizator",
    "description": "Creează o relație de urmărire cu un alt utilizator și generează o notificare",
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "ID-ul utilizatorului de urmărit"
      }
    ],
    "responses": {
      "201": {
        "description": "Utilizatorul a fost urmărit cu succes",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "400": {
        "description": "Eroare de validare (nu poți să te urmărești pe tine însuți sau utilizatorul este deja urmărit)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
},
"/api/interactions/is-following/{userId}": {
  "get": {
    "summary": "Verifică dacă urmărești un utilizator",
    "description": "Returnează statusul de urmărire pentru un utilizator specific",
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "ID-ul utilizatorului"
      }
    ],
    "responses": {
      "200": {
        "description": "Status de urmărire",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "isFollowing": { "type": "boolean" }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
},
"/api/interactions/notifications": {
  "get": {
    "summary": "Obține notificările utilizatorului",
    "description": "Returnează ultimele 50 de notificări ale utilizatorului autentificat, ordonate descrescător după data creării",
    "security": [{ "bearerAuth": [] }],
    "responses": {
      "200": {
        "description": "Lista notificărilor",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "userId": { "type": "integer" },
                  "message": { "type": "string" },
                  "isRead": { "type": "boolean" },
                  "createdAt": { "type": "string", "format": "date-time" },
                  "updatedAt": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
},
"/api/interactions/notifications/unread-count": {
  "get": {
    "summary": "Obține numărul de notificări necitite",
    "description": "Returnează numărul total de notificări necitite ale utilizatorului autentificat",
    "security": [{ "bearerAuth": [] }],
    "responses": {
      "200": {
        "description": "Numărul de notificări necitite",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "count": { "type": "integer" }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
},
"/api/interactions/notifications/mark-read": {
  "put": {
    "summary": "Marchează notificările ca citite",
    "description": "Marchează notificări specifice sau toate notificările ca citite",
    "security": [{ "bearerAuth": [] }],
    "requestBody": {
      "required": false,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "notificationIds": {
                "type": "array",
                "items": { "type": "integer" },
                "description": "Array de ID-uri de notificări. Dacă nu este furnizat, toate notificările vor fi marcate ca citite"
              }
            }
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Notificările au fost marcate ca citite",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
},
"/api/interactions/notifications/{id}": {
  "delete": {
    "summary": "Șterge o notificare",
    "description": "Șterge o notificare specifică a utilizatorului autentificat",
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "ID-ul notificării"
      }
    ],
    "responses": {
      "200": {
        "description": "Notificarea a fost ștearsă",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "404": {
        "description": "Notificarea nu a fost găsită",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
},
"/api/interactions/unfollow/{userId}": {
  "delete": {
    "summary": "Anulează urmărirea unui utilizator",
    "description": "Elimină relația de urmărire cu un alt utilizator",
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "ID-ul utilizatorului de anulat din urmărire"
      }
    ],
    "responses": {
      "200": {
        "description": "Utilizatorul a fost anulat din urmărire cu succes",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "400": {
        "description": "Nu urmărești acest utilizator",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
},
"/api/interactions/followers/count": {
  "get": {
    "summary": "Obține numărul de urmăritori ai utilizatorului curent",
    "description": "Returnează numărul total de urmăritori pentru utilizatorul autentificat",
    "security": [{ "bearerAuth": [] }],
    "responses": {
      "200": {
        "description": "Numărul de urmăritori",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "count": { "type": "integer" }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
},
"/api/interactions/following/count": {
  "get": {
    "summary": "Obține numărul de utilizatori urmăriți de utilizatorul curent",
    "description": "Returnează numărul total de utilizatori pe care îi urmărește utilizatorul autentificat",
    "security": [{ "bearerAuth": [] }],
    "responses": {
      "200": {
        "description": "Numărul de utilizatori urmăriți",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "count": { "type": "integer" }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
},
"/api/interactions/{userId}/followers/count": {
  "get": {
    "summary": "Obține numărul de urmăritori ai unui utilizator specific",
    "description": "Returnează numărul total de urmăritori pentru un utilizator specific",
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "ID-ul utilizatorului"
      }
    ],
    "responses": {
      "200": {
        "description": "Numărul de urmăritori",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "count": { "type": "integer" }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
},
"/api/interactions/{userId}/following/count": {
  "get": {
    "summary": "Obține numărul de utilizatori urmăriți de un utilizator specific",
    "description": "Returnează numărul total de utilizatori pe care îi urmărește un utilizator specific",
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "ID-ul utilizatorului"
      }
    ],
    "responses": {
      "200": {
        "description": "Numărul de utilizatori urmăriți",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "count": { "type": "integer" }
              }
            }
          }
        }
      },
      "500": {
        "description": "Eroare server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
}
  }
}